requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
APScheduler
plyer
sentence-transformers
chromadb
